added forward-looking to optimizer
