+++ assets/js/orders.js
@@ -1,9 +1,9 @@
 module.exports = [
-    { d: 225, l: 77, q: 10 },
+    { d: 225, l: 77, q: 5 },
     // { d: 2, l: 93, q: 1 },
     // { d: 2, l: 74, q: 1 },
     { d: 225, l: 69.25, q: 1 },
-    { d: 225, l: 77, q: 12 },
+    { d: 225, l: 77, q: 8 },
     { d: 225, l: 52, q: 2 },
     { d: 225, l: 80, q: 2 },
     // { d: 2, l: 60, q: 1 },
